requirements.txt
pytz==2022.6
click==8.1.3
requests==2.28.1
celery==5.4.0
colorama==0.4.6
fastapi==0.115.11
httpx==0.28.1
langchain==0.3.20
langchain_core==0.3.43
langchain_openai==0.3.8
loguru==0.7.3
openai==1.65.4
Pillow==11.1.0
pydantic==2.10.6
pydantic_core
pydantic_settings==2.8.1
pytest==8.3.5
PyYAML==6.0.2
twikit==2.3.3
uvicorn==0.34.0
websocket_client==1.8.0
mypy==1.15.0
ollama==0.4.7
pandas==2.2.3
colorama==0.4.6
twikit==2.3.3
python-box==7.3.2
openai==1.65.4
mcp==1.6.0
faiss-cpu==1.10.0
psutil==6.0.0
tenacity==8.2.3
redis==5.2.1
gevent==25.4.2
googlesearch-python==1.3.0
pymysql==1.1.1
twine==6.1.0
